# docs/api-ledger/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
...
      "CompletePaymentResource" : {
        "properties" : {
          "transactionReference" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CreatePaymentResource" : {
        "properties" : {
          "amount" : {
            "format" : "double",
            "type" : "number"
          },
          "companyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "orderId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "paymentMethod" : {
            "enum" : [ "BANK_TRANSFER", "CREDIT_CARD", "DEBIT_CARD", "CASH" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PaymentResource" : {
        "properties" : {
          "amount" : {
            "format" : "double",
            "type" : "number"
          },
          "companyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "orderId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "paidAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "paymentMethod" : {
            "enum" : [ "BANK_TRANSFER", "CREDIT_CARD", "DEBIT_CARD", "CASH" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING", "COMPLETED", "FAILED", "REFUNDED" ],
            "type" : "string"
          },
          "transactionReference" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ProviderPaymentResource" : {
        "properties" : {
          "amount" : {
            "format" : "double",
            "type" : "number"
          },
          "buyerCompanyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "buyerName" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency" : {
            "description" : "Moneda no persistida actualmente; null, nunca inferida",
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "orderId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "paidAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "paymentMethod" : {
            "enum" : [ "BANK_TRANSFER", "CREDIT_CARD", "DEBIT_CARD", "CASH" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING", "COMPLETED", "FAILED", "REFUNDED" ],
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Token JWT Bearer para autenticar solicitudes a la API",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "support@fuelguard.app",
      "name" : "Soporte de FuelGuard",
      "url" : "https://fuelguard.app/support"
    },
    "description" : "API de FuelGuard para gestionar abastecimiento, flota, entregas, inventario, telemetría IoT y pagos por organización.",
    "license" : {
      "name" : "Apache 2.0 (licencia)",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "FuelGuard Platform",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/payments" : {
      "get" : {
        "description" : "Devuelve todos los pagos registrados; requiere autoridad administrativa ROLE_ADMIN.",
        "operationId" : "getAllPayments",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PaymentResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Se devuelve la lista de pagos."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PaymentResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "El usuario no cuenta con autoridad ROLE_ADMIN."
          }
        },
        "summary" : "Listar todos los pagos",
        "tags" : [ "Pagos" ]
      },
      "post" : {
        "description" : "Registra el pago de una orden después de verificar la empresa propietaria, la existencia de la orden y la coincidencia del importe total.",
        "operationId" : "createPayment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePaymentResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Pago registrado."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Falta la orden o la empresa, o el importe no coincide con el total de la orden."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "La empresa compradora indicada no pertenece al usuario autenticado."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "La orden no existe o no pertenece a la empresa indicada."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "La orden ya tiene un pago registrado."
          }
        },
        "summary" : "Registrar pago",
        "tags" : [ "Pagos" ]
      }
    },
    "/api/payments/company/{companyId}" : {
      "get" : {
        "description" : "Devuelve los pagos de la empresa indicada si coincide con la empresa del usuario autenticado.",
        "operationId" : "getPaymentsByCompany",
        "parameters" : [ {
          "in" : "path",
          "name" : "companyId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PaymentResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Se devuelve la lista de pagos de la empresa."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PaymentResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "La empresa solicitada no pertenece al usuario autenticado."
          }
        },
        "summary" : "Listar pagos por empresa",
        "tags" : [ "Pagos" ]
      }
    },
    "/api/payments/order/{orderId}" : {
      "get" : {
        "description" : "Devuelve el pago asociado a la orden si el usuario pertenece a la empresa compradora o al distribuidor.",
        "operationId" : "getPaymentByOrder",
        "parameters" : [ {
          "in" : "path",
          "name" : "orderId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResource"
                }
              }
            },
            "description" : "Pago devuelto."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResource"
                }
              }
            },
            "description" : "La orden no tiene pago o este no es visible para el usuario."
          }
        },
        "summary" : "Consultar pago de una orden",
        "tags" : [ "Pagos" ]
      }
    },
    "/api/payments/provider/{providerId}" : {
      "get" : {
        "description" : "Solo el proveedor autenticado propietario. Una consulta por proveedor, con comprador y orden. status es PaymentStatus; from/to son instantes ISO inclusivos sobre createdAt. Orden descendente por fecha e id. Moneda null porque no se persiste en pagos ni órdenes.",
        "operationId" : "getPaymentsByProvider",
        "parameters" : [ {
          "in" : "path",
          "name" : "providerId",
          "required" : true,
          "schema" : {
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "PENDING", "COMPLETED", "FAILED", "REFUNDED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ProviderPaymentResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Pagos propios o []"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ProviderPaymentResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Id, estado, fecha o período inválido"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ProviderPaymentResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Sin rol proveedor o proveedor ajeno"
          }
        },
        "summary" : "Listar pagos del distribuidor",
        "tags" : [ "Pagos" ]
      }
    },
    "/api/payments/{paymentId}" : {
      "get" : {
        "description" : "Devuelve el pago si el usuario pertenece a la empresa compradora o al distribuidor de la orden.",
        "operationId" : "getPaymentById",
        "parameters" : [ {
          "in" : "path",
          "name" : "paymentId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResource"
                }
              }
            },
            "description" : "Pago devuelto."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentResource"
                }
              }
            },
            "description" : "El pago no existe o no es visible para el usuario."
          }
        },
        "summary" : "Consultar pago por identificador",
        "tags" : [ "Pagos" ]
      }
    },
    "/api/payments/{paymentId}/complete" : {
      "post" : {
        "description" : "Marca el pago como completado con la referencia de transacción y actualiza la orden a pagada.",
        "operationId" : "completePayment",
        "parameters" : [ {
          "in" : "path",
          "name" : "paymentId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompletePaymentResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Pago completado y orden marcada como pagada."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El pago o su orden no existe, o el usuario no puede consultarlo."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "La orden no puede marcarse como pagada, por ejemplo, porque fue cancelada."
          }
        },
        "summary" : "Completar pago",
        "tags" : [ "Pagos" ]
      }
    },
    "/api/payments/{paymentId}/refund" : {
      "post" : {
        "description" : "Marca como reembolsado un pago completado; repetir la operación conserva ese estado.",
        "operationId" : "refundPayment",
        "parameters" : [ {
          "in" : "path",
          "name" : "paymentId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Pago reembolsado."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El pago no existe o el usuario no puede consultarlo."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El pago no está completado (PENDING o FAILED) y no puede reembolsarse."
          }
        },
        "summary" : "Reembolsar pago",
        "tags" : [ "Pagos" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Servidor de la API",
    "url" : "http://localhost:8080"
  } ],
...
}